Add files via upload
Mental illness is a health problem that undoubtedly impacts emotions, 
reasoning, and social interaction of a person. After Covid 19, especially the PG students on 
the brink of making their careers faced these issues widely. This makes them vulnerable to 
various psychiatric conditions.By considering all these things I decided to do our project work on this subject. 
To achieve our claim, it is important to collect the response of the students from Shivaji 
university, Kolhapur. We collected the data from post-graduation students of this university.